"Like with like" - move a few functions to better places in the code (they are placed strangely)
A few functions such as ipcmp(), subnet_expand(), and check_subnets_overlap()  are in illogical places - away from all the other ip comparison and subnet basic functions and in the middle of alias handling and interface enumeration. 

No change to functional code, just moving to earlier in the file (next to other IP compare and subnet functions) for ease of future contributors.