Firewall Log does not display logged IGMP packets
If IGMP packets are logged (either pass or block) then parse_filter_line did not set their src and dst IP.
Later in the subroutine, it zapped the filter line because it did not have a src and dst.
This fixes it. Now the IGMP lines in /var/log/filter.log appear on the Firewall Log GUI.