DNS Forwarder Host Override Edit make the pencil symbols appear
In most places in the GUI a little pencil symbol appears before text input fields. For example it already appears for the Host, Domain and Description fields in this very form when you "Enter additional names for this host."
But in the main 4 fields of this form, the pencil does not display.
This fixes it to be the same as elsewhere.
I guess this is the "standard" supposed to be used. If so, then I will find the other places where class="formfld unknown" has the "unknown" missing.