Pencil symbols
These are places in the GUI where the cursor sits not in the far left
side of the input box and there is odd-looking white space to the left
of the cursor. Normally there would be a little input graphic in the
white space to the left of the cursor (a pencil, a computer screen, a
lock symbol...)
This change makes the pencil be displayed in all those places.