Fix SNMP Bind Interface list in master
The format of the array returned by get_possible_listen_ips() was changed.
This use of the returned array was not quite changed correctly. This fixes it so the drop-down list of SNMP Bind Interfaces is displayed correctly.
Note that the change to get_possible_listen_ips() (which also moved the function from system.inc to interfaces.inc) and associated effects is only in master. It has not been applied to RELENG_2_2 yet. So this fix is only needed in master.
I noticed this when using various files from master on a test system.