Do not allow VLAN tag zero
At the moment you can make a VLAN with tag 0. The input validation does not catch it because when $_POST['tag'] = "0" that evaluates to false by PHP.
There is no need for that check here, because do_input_validation already checks that $_POST['tag'] is non-empty, and we really always want the other numericint and limit checks here.