Domain override with multiple authoritative DNS servers
Tell users that this is possible in DNS Resolver and how to achieve it. The code in unbound.inc already supports it and works.
I had asked for this in Redmine feature request #4350 and when I went to look at coding to implement it I found code that already did it.
So IMHO it is worth telling users.