status_rd_graph.php updated
This page has been extensively updated, but it could do with bing
revised further at some time in the future.

For now it works slightly better than the original, looks reasonable
and has been updated to the project display system.

I’d like to merge it now, with the option to revisit it later.