Independent Dashboard Theme
With the current 'Theme' design, the chosen theme is used in all pages.

This work-around will allow the setting of a 'theme' for the Dashboard
Independent of the Theme for the balance of the pfSense Pages.

Use case would be setting the 'Dashboard' to a Wide-Screen theme, and
leaving all other pages in a Standard Width theme.

When the   index.php   file is loaded, if the "Dashboard Theme" is
defined with a theme that is different than the "Default Theme", the
"Default Theme" setting is captured into $g['theme_revert']. The new
"Dashboard Theme" is now used for the Dashboard. All other pages will
use the "Default Theme" setting.

If the user applies any changes in the Dashboard Widget, it will save
the "Dashboard Theme" as the "Default Theme" and this will cause all
future pages to use the "Dashboard Theme". To over come this issue, code
is added to the php function    write_config()    to revert the "Theme"
back to its "Default" before saving the changes to the config.xml file.

Future designs of Themes should allow for independent Theme
customization per page to better utilize
Wide-Screen/Standard-Width-Screen settings.