diag_logs - log file consolidation
A GET argument has been added to dig_logs.php so that the log file to
displayed can be specified.

Minor edits to the file allow all log files to be displayed correctly

The tab array has been updated in this file, and in:

    diag_logs_settings.php
    diag_logs_vpn.php
    diag_logs_filter.php
    diag_logs_filter_dynamic.php
    diag_logs_filter_summary.php

The last four of these files have not yet been converted to bootstrap,
but the tab arrays are now compatible with this change.

The previous PR for diag_logs_settings will be deleted.

The following files are no longer required and have been removed:

    diag_logs_auth.php
    diag_logs_dhcp.php
    diag_logs_gateways.php
    diag_logs_ipsec.php
    diag_logs_ntpd.php
    diag_logs_openvpn.php
    diag_logs_ppp.php
    diag_logs_relayd.php
    diag_logs_resolver.php
    diag_logs_routing.php
    diag_logs_wireless.php