diag_logs_filter_summary conversion complete
Ready for review

The previous charing system is not compatible with JQuery (because it
makes use of its own $() macros and converting the code would be
horrible. It is not responsive or attractive anyway.

The library has been removed and replaced by d3pie.js. The page has
been re-written to use the new library.

Seems to look very nice and the code is quite clean.