Minor changes to firewall_shaper_layer7
The tabbing of this code is not so good, so it is difficult to see what is going on. I will format that in a later pull request.
These things look unbalanced but actually were not at run time. These changes make it easier to check the code in an editor that can show the balanced "table" "/table", "selected" "/selected".
1) Take the "table" start declaration out of the output form var that has been built up. Put it in-line in the code so it can be seen to match the "/table" now at line 575. $output var is no longer needed.
2) Remove 3 "/select" that were each inside an "if" statement. Put a single "/select" at the end. This makes it easy to see the matching "select" and "/select" and also ensures there will always be a "/select" written.

I thought these things were worth doing separate to code-style, as they are "real" changes. The L7 form still works in my testing.