Setup Wizard can result in invalid LAN DHCP pool calculation
1) consider where the LAN IP is in the subnet range and then put the
DHCP pool in the biggest remaining segment, either above or below.
2) Check the size of the available segment. If it is reasonably big then
leave some space at either end of the segment, like the old code was
doing. Otherwise give all the space to the pool.
3) Do not allow subnet mask 32 - I can't think of a use case for LAN to
have a /32 subnet mask, it kind of breaks the whole concept of LAN.
4) Provide more detailed separate messages if the user tries to use the
network address or broadcast address as the LAN IP.