Improve setup wizard host name check
Redmine #4712
It seems good enough to make the regex strings here be "reasonable". The full checks are done after pressing Next and the correct routines are called that do an exhaustive check. There seems not much point in trying to re-engineer all that here also.
Odd things like "-hostname" and "hostname-" would be allowed through here but are caught by the full validation check.
"." and "|" were being allowed in this regex - no idea why!