Add description as a display option on Traffic Graph
This is handy at sites where lots of the LAN clients have static-mapped
DHCP IP addresses. Depending on the site host naming conventions, host
names can be a bit obtuse - may not tell you where the client device
might actually be in the building. We put other useful stuff in the
description - "Jane Doe - Reception".
This enhancement allows the user of Traffic Graph to select
"Description" in the "Display" dropdown. Then, for IP addresses that are
static mapped, the description from the config is shown, rather than the
host name.
When a client device is noticed hogging bandwidth, it is easy to go
straight to "Jane Doe - Reception" and ask why they are doing some huge
download.
Might be useful for for others?