Say what is happening when reinstalling package GUI XML
At the confirmation dialog after pressing the "Reinstall XML" button, the text does not distinguish between having pressed "Reinstall the whole package" and "Reinstall the GUI/XML". It would be nice if the text of this confirmation allowed the user to be confident about which button they had just pushed, before confirming the action.
Note: This stuff has no gettext() wrappers - but that can be fixed later, not get mixed up in this.