Display monitor IP on Gateways widget
This change adds a setting for the Gateways dashboard widget so the user
can choose to display the Gateway IP, Monitor IP or both.
If "both" is chosen and the Gateway IP is the Monitor IP, then only the
Gateway IP is shown - i.e. the same IP address is not repeated on the
widget display.
If "both" is chosen and the Monitor IP is different to the Gateway IP
then the Monitor IP is shown in () brackets after the Gateway IP.
If "Monitor IP" is chosen and there is no special Monitor IP defined,
then the Gateway IP is displayed (which is also the Monitor IP).
If "Gateway IP" is chosen then the widget behaves as it does now.
"Gateway IP" is the default.
I find this handy because the Gateways widget reports RTT (latency) and
loss figures that are actually for pings to the Monitor IP. So it seems
useful to be able to display the Monitor IP in the widget.