Add DNS host override descriptions
since we can sometimes provide a useful description from that config
data also.
Fill the $iplookup array with host or FQDN data if description is blank
or host or FQDN was requested. Then we can use $iplookup in all cases
where we have local data. It simplifies some logic a bit.