Fix references to Load Balancer Virtual Server redirect_mode
When adding a Virtual Server, if you press Save with blank fields, the validation does not show. That was because the switch statement selecting the fields to validate had wrong case values. Actually the "mode" here is locked to "redirect_mode" (later code to allow the mode to be selected by the user is commented out - not implemented.
I fixed the reqdfieldsn array also for relay_mode case, even though it is never used. It looked dodgy the way it was.