Unbound GUI ("Advanced" -> "Custom options")
A user configuring Unbound will see a tab for "advanced" options, but also a custom config input labelled "advanced". Odd to say that this is the input box for "advanced" options when there's also a tab for "advanced". Retitle the input box to "Custom options" which is more accurate. 

(OK it's minor but shouldn't have 2 Unbound entrypoints *both* called "Advanced".....)