Validate unbound.conf changes before save+resync
Currently the only way bad custom options can be detected is to save them, but even so, no error detection occurs, therefore new options saved, but unbound fails to restart and the old config has been overwritten already and can't be reverted.

Having modified unbound_generate_config() to allow testing-only, we can validate a proposed config change before irrevocably overwriting the current unbound.conf, and notify the user if it won't work (eg if they had an error in their custom options).

This is ugly code unfortunately hence tagged with "FIXME". It works, but it relies on changing $config, then changing back if validation failed, rather than constructing the new (proposed) $config['unbound'] in a temporary array, and only overwriting $config if validation *succeeded*, which would be better. I've done this to avoid too-extensive changes to the codebase. Also unbound-checkconf doesn't specify the errors found, so if an error is returned, we have to find what it was ourselves. To be somewhat helpful the code re-tries it without the custom options changes, on the basis that the most helpful thing is to identify whether the freeform input was the apparent sole problem. Again this could be done better but for now it's at least more helpful than present code.

The effect of these changes is that a non-valid new config won't so easily overwrite the current config, the resolver won't be told to resync until the new conf is valid, and in the event that the error is purely in the custom options this will be detected and the user told. So the resolver and config changes to it, should be less easily upset by accidental bad input.